+++ Spring/post-example/src/main/java/com/practice/postexample/dto/articleComment/ArticleCommentDto.java
@@ -1,37 +1,36 @@
-package com.practice.postexample.dto;
+package com.practice.postexample.dto.articleComment;
 
-import com.practice.postexample.domain.Article;
+import com.practice.postexample.domain.ArticleComment;
+import com.practice.postexample.dto.userAccount.UserAccountDto;
+import com.practice.postexample.dto.article.ArticleDto;
 
 import java.time.LocalDateTime;
 
-public record ArticleDto(
+public record ArticleCommentDto(
         Long id,
+        ArticleDto articleDto,
         UserAccountDto userAccountDto,
-        String title,
         String content,
-        String hashtag,
         LocalDateTime createdAt,
         String createdBy,
         LocalDateTime modifiedAt,
         String modifiedBy
 ) {
-    public static ArticleDto of(
+    public static ArticleCommentDto of(
             Long id,
+            ArticleDto articleDto,
             UserAccountDto userAccountDto,
-            String title,
             String content,
-            String hashtag,
             LocalDateTime createdAt,
             String createdBy,
             LocalDateTime modifiedAt,
             String modifiedBy
     ) {
-        return new ArticleDto(
+        return new ArticleCommentDto(
                 id,
+                articleDto,
                 userAccountDto,
-                title,
                 content,
-                hashtag,
                 createdAt,
                 createdBy,
                 modifiedAt,
@@ -39,13 +38,12 @@ public record ArticleDto(
         );
     }
 
-    public static ArticleDto from(Article entity) {
-        return new ArticleDto(
+    public static ArticleCommentDto from(ArticleComment entity) {
+        return new ArticleCommentDto(
                 entity.getId(),
+                ArticleDto.from(entity.getArticle()),
                 UserAccountDto.from(entity.getUserAccount()),
-                entity.getTitle(),
                 entity.getContent(),
-                entity.getHashtag(),
                 entity.getCreatedAt(),
                 entity.getCreatedBy(),
                 entity.getModifiedAt(),
@@ -53,12 +51,11 @@ public record ArticleDto(
         );
     }
 
-    public Article toEntity() {
-        return Article.builder()
+    public ArticleComment toEntity() {
+        return ArticleComment.builder()
+                .article(articleDto.toEntity())
                 .userAccount(userAccountDto.toEntity())
-                .title(title)
                 .content(content)
-                .hashtag(hashtag)
                 .build();
     }
 }
